Remove redundant parts
